+++ tests/testthat/test-apply_mapping.R
@@ -109,26 +109,6 @@ describe("apply_mapping", {
     expect_equal(result_df, expected_df)
   })
 
-  it("removes concentration duplicates only if exact", {
-    test_df <- rbind(expected_df, expected_df %>% mutate(AVAL = 1))
-    filtered_df <- test_df %>%
-      filter(!duplicated(paste(USUBJID, AVAL))) %>%
-      arrange(USUBJID, AVAL)
-
-    # add label attr to NCA profile
-    var_labels(filtered_df) <- var_labels(expected_df)
-
-    mapping <- as.list(setNames(names(test_df), names(test_df)))
-    expect_warning(
-      apply_mapping(dataset = test_df, mapping = mapping, desired_order = desired_order),
-      "Duplicate concentration data detected and filtered"
-    )
-    result_df <- suppressWarnings(
-      apply_mapping(dataset = test_df, mapping = mapping, desired_order = desired_order)
-    )
-    expect_equal(result_df, filtered_df)
-  })
-
   it("allows duplicated mappings", {
     mapping["PARAM"] <- mapping["DOSETRT"]
     result_df <- apply_mapping(dataset = test_df, mapping = mapping, desired_order = desired_order)